reorganize quest shapes
